tsm1: move cache entry to its own file
